Put the switch service-launching logic back
